Extract HloExecutionProfiler into its own file.
This is in preparation of passing it on to the Thunks, so that we can profile
HloInstructions within a while loop.

PiperOrigin-RevId: 200532394
